backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg[binary]==3.2.9
pydantic==2.7.4
pydantic-settings==2.3.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.6
rq==1.16.2
httpx==0.27.0
itsdangerous==2.2.0
orjson==3.10.5
sse-starlette==2.1.0 